refactor: move MyChart to lib/
