+++ test/hello.test.ts
@@ -1,0 +1,11 @@
+import { HelloChart } from "../lib/hello-chart";
+import { Testing } from "cdk8s";
+
+describe("Placeholder", () => {
+  test("Empty", () => {
+    const app = Testing.app();
+    const chart = new HelloChart(app, "test-chart");
+    const results = Testing.synth(chart);
+    expect(results).toMatchSnapshot();
+  });
+});
